checkpath: fix toctou issues
This walks the directory path to the file we are going to create to make
sure that when we create the file and change the ownership and permissions
we are working on the same file.
Also, symbolic links in the path need to be owned by root or the current
user.

On non-linux platforms, we no longer follow non-terminal symbolic links
by default. If you need to do that, add the -s option on the checkpath
command line.

This is for #201.